ordering-api/pyproject.toml
[tool.poetry]
name = "gtf_order_api"
version = "0.1.0"
description = ""
authors = ["Semicolon1647 <Semicolon1647@proton.me>"]
license = "AGPL-3.0"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
django-stubs = "^1.10.1"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "myproject.settings"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

